frontend 2 updated

+++ src/app/student/student-navbar/student-navbar.component.ts
@@ -18,9 +18,15 @@ export class StudentNavbarComponent implements OnInit {
 
   logout() {
 
-    sessionStorage.clear();
+    const confirmLogout = confirm("Are you sure you want to logout?");
 
-    this.router.navigate(['/student-login']);
+    if (confirmLogout) {
+
+      sessionStorage.clear();
+
+      this.router.navigateByUrl('/', { replaceUrl: true });
+
+    }
 
   }
